Rename Change\Comment to ChangeComment
