+++ app/Actions/RespondToBookingAction.php
@@ -6,6 +6,7 @@ use App\Enums\AttendeeStatus;
 use App\Models\Booking;
 use App\Models\Change;
 use App\Models\ChangeAttendee;
+use App\Models\ChangeComment;
 use App\Models\User;
 use InvalidArgumentException;
 
@@ -63,11 +64,11 @@ class RespondToBookingAction
             }
             $change->attendees()->save($change_attendee);
         } else if (!empty($comment) && ($attendance?->comment != $comment)) {
-            $change = new Change();
+            $change = new Change;
             $change->author_id = $author_id;
             $this->booking->changes()->save($change);
 
-            $change_comment = new Change\Comment;
+            $change_comment = new ChangeComment;
             $change_comment->body = $comment;
             $change->comments()->save($change_comment);
         }
